Python 3 compat: handle removal of the "long" type
Reviewed-on: https://trac.frm2.tum.de/review/4328
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>

+++ lib/nicos/devices/tango.py
@@ -359,7 +359,7 @@ class DigitalOutput(PyTangoDevice, Moveable):
         return self._dev.value
 
     def doStart(self, value):
-        self._dev.value = long(value)
+        self._dev.value = int(value)
 
 
 class StringIO(PyTangoDevice, Device):
